requirements.txt
discord.py
python-dotenv
lunardate
requests
pytz
